Move token to own module
