+++ parser/src/lib.rs
@@ -1,4 +1,5 @@
-mod ast;
+mod token;
 mod lexer;
+mod ast;
 pub mod parser;
 mod token_reader;
